Rename to distinguish from parser
