+++ src/features/suggestions/parser/DecomposedAmount.ts
